// frontend/generated-sources/backend-api/models/CameraBounds.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Point2d } from '../models/Point2d';
import { HttpFile } from '../http/http';

export class CameraBounds {
    'cameraBoundBottomLeft'?: Point2d;
    'cameraBoundTopRight'?: Point2d;
    'cameraBoundTopLeft'?: Point2d;
    'cameraBoundBottomRight'?: Point2d;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "cameraBoundBottomLeft",
            "baseName": "cameraBoundBottomLeft",
            "type": "Point2d",
            "format": ""
        },
        {
            "name": "cameraBoundTopRight",
            "baseName": "cameraBoundTopRight",
            "type": "Point2d",
            "format": ""
        },
        {
            "name": "cameraBoundTopLeft",
            "baseName": "cameraBoundTopLeft",
            "type": "Point2d",
            "format": ""
        },
        {
            "name": "cameraBoundBottomRight",
            "baseName": "cameraBoundBottomRight",
            "type": "Point2d",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CameraBounds.attributeTypeMap;
    }

    public constructor() {
    }
}