// frontend/generated-sources/backend-api/models/Force.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ForceFlags } from '../models/ForceFlags';
import { PlayerBitmap } from '../models/PlayerBitmap';
import { W3str } from '../models/W3str';
import { HttpFile } from '../http/http';

export class Force {
    'flags'?: ForceFlags;
    'players'?: PlayerBitmap;
    'name'?: W3str;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "flags",
            "baseName": "flags",
            "type": "ForceFlags",
            "format": ""
        },
        {
            "name": "players",
            "baseName": "players",
            "type": "PlayerBitmap",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "W3str",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Force.attributeTypeMap;
    }

    public constructor() {
    }
}