// frontend/generated-sources/backend-api/models/ForceFlags.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ForceFlags {
    'flag7'?: boolean;
    'flag6'?: boolean;
    'flag5'?: boolean;
    'shareAdvancedUnitControl'?: boolean;
    'shareUnitControl'?: boolean;
    'shareVision'?: boolean;
    'alliedVictory'?: boolean;
    'allied'?: boolean;
    'rest'?: Array<number>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "flag7",
            "baseName": "flag7",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "flag6",
            "baseName": "flag6",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "flag5",
            "baseName": "flag5",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "shareAdvancedUnitControl",
            "baseName": "shareAdvancedUnitControl",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "shareUnitControl",
            "baseName": "shareUnitControl",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "shareVision",
            "baseName": "shareVision",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "alliedVictory",
            "baseName": "alliedVictory",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "allied",
            "baseName": "allied",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "rest",
            "baseName": "rest",
            "type": "Array<number>",
            "format": "int32"
        }    ];

    static getAttributeTypeMap() {
        return ForceFlags.attributeTypeMap;
    }

    public constructor() {
    }
}