// frontend/generated-sources/backend-api/models/ForcesChunk.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Force } from '../models/Force';
import { HttpFile } from '../http/http';

export class ForcesChunk {
    'numForce'?: number;
    'force'?: Array<Force>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "numForce",
            "baseName": "numForce",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "force",
            "baseName": "force",
            "type": "Array<Force>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ForcesChunk.attributeTypeMap;
    }

    public constructor() {
    }
}