// frontend/generated-sources/backend-api/models/GameVersion.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class GameVersion {
    'major'?: number;
    'minor'?: number;
    'revision'?: number;
    'build'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "major",
            "baseName": "major",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "minor",
            "baseName": "minor",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "revision",
            "baseName": "revision",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "build",
            "baseName": "build",
            "type": "number",
            "format": "int64"
        }    ];

    static getAttributeTypeMap() {
        return GameVersion.attributeTypeMap;
    }

    public constructor() {
    }
}