// frontend/generated-sources/backend-api/models/Item.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { W3id } from '../models/W3id';
import { HttpFile } from '../http/http';

export class Item {
    'chance'?: number;
    'id'?: W3id;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "chance",
            "baseName": "chance",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "W3id",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Item.attributeTypeMap;
    }

    public constructor() {
    }
}