// frontend/generated-sources/backend-api/models/ItemSet.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Item } from '../models/Item';
import { HttpFile } from '../http/http';

export class ItemSet {
    'numItem'?: number;
    'item'?: Array<Item>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "numItem",
            "baseName": "numItem",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "item",
            "baseName": "item",
            "type": "Array<Item>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ItemSet.attributeTypeMap;
    }

    public constructor() {
    }
}