// frontend/generated-sources/backend-api/models/Margins.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Margins {
    'left'?: number;
    'right'?: number;
    'bottom'?: number;
    'top'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "left",
            "baseName": "left",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "right",
            "baseName": "right",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "bottom",
            "baseName": "bottom",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "top",
            "baseName": "top",
            "type": "number",
            "format": "int64"
        }    ];

    static getAttributeTypeMap() {
        return Margins.attributeTypeMap;
    }

    public constructor() {
    }
}