// frontend/generated-sources/backend-api/models/Player.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PlayerBitmap } from '../models/PlayerBitmap';
import { Point2d } from '../models/Point2d';
import { W3str } from '../models/W3str';
import { HttpFile } from '../http/http';

export class Player {
    'num'?: number;
    'controller'?: PlayerControllerEnum;
    'race'?: PlayerRaceEnum;
    'fixedPosition'?: number;
    'name'?: W3str;
    'position'?: Point2d;
    'allyLowPriority'?: PlayerBitmap;
    'allyHighPriority'?: PlayerBitmap;
    'enemyLowPriority'?: PlayerBitmap;
    'enemyHighPriority'?: PlayerBitmap;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "num",
            "baseName": "num",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "controller",
            "baseName": "controller",
            "type": "PlayerControllerEnum",
            "format": ""
        },
        {
            "name": "race",
            "baseName": "race",
            "type": "PlayerRaceEnum",
            "format": ""
        },
        {
            "name": "fixedPosition",
            "baseName": "fixedPosition",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "W3str",
            "format": ""
        },
        {
            "name": "position",
            "baseName": "position",
            "type": "Point2d",
            "format": ""
        },
        {
            "name": "allyLowPriority",
            "baseName": "allyLowPriority",
            "type": "PlayerBitmap",
            "format": ""
        },
        {
            "name": "allyHighPriority",
            "baseName": "allyHighPriority",
            "type": "PlayerBitmap",
            "format": ""
        },
        {
            "name": "enemyLowPriority",
            "baseName": "enemyLowPriority",
            "type": "PlayerBitmap",
            "format": ""
        },
        {
            "name": "enemyHighPriority",
            "baseName": "enemyHighPriority",
            "type": "PlayerBitmap",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Player.attributeTypeMap;
    }

    public constructor() {
    }
}


export type PlayerControllerEnum = "NONE" | "HUMAN" | "CPU" | "NEUTRAL" | "RESCUABLE" ;
export type PlayerRaceEnum = "SELECTABLE" | "HUMAN" | "ORC" | "UNDEAD" | "NIGHT_ELF" ;