// frontend/generated-sources/backend-api/models/PlayerBitmap.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class PlayerBitmap {
    'isMember'?: Array<boolean>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "isMember",
            "baseName": "isMember",
            "type": "Array<boolean>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return PlayerBitmap.attributeTypeMap;
    }

    public constructor() {
    }
}