// frontend/generated-sources/backend-api/models/Point2d.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Point2d {
    'x'?: number;
    'y'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "x",
            "baseName": "x",
            "type": "number",
            "format": "float"
        },
        {
            "name": "y",
            "baseName": "y",
            "type": "number",
            "format": "float"
        }    ];

    static getAttributeTypeMap() {
        return Point2d.attributeTypeMap;
    }

    public constructor() {
    }
}